BlackoilModelEbos: put Nldd solver in separate class
